PR comments fixed
