+++ src/scgpt/cross_check_genes/config.vsh.yaml
@@ -1,5 +1,5 @@
 functionality:
-  name: cross_check
+  name: cross_check_genes
   namespace: "scgpt"
   description: |
     Cross-check genes with pre-trained model.
@@ -21,20 +21,20 @@ functionality:
           type: string
           default: "rna"
           required: false
-        - name: "--model_dir"
+        - name: "--vocab_file"
           type: file
           direction: input
           required: true
-          example: foundational_model/scgpt
+          example: resources_test/scgpt/vocab.json
           description: |
-            The directory containing the model files. Must contain 'vocab.json', 'model.pt' and 'args.json' files.
+            Model vocabulary file directory.
         - name: "--gene_name_layer"
           type: string
           example: "gene_name"
           required: false
           default: "gene_name"
           description: |
-            The name of the layer containing gene names.
+            The name of the adata.var column containing gene names.
 
     - name: Outputs
       arguments:
@@ -54,11 +54,6 @@ functionality:
           required: false
           description: |
             The padding token used in the model.
-        - name: "--load_model_vocab"
-          type: boolean
-          default: true
-          description: |
-            Whether to load the vocabulary from the model or use pytorch vocab.
         - name: "--ori_batch_layer_name"
           type: string
           default: "sample"
@@ -74,6 +69,7 @@ functionality:
     - type: python_script
       path: test.py
     - path: /resources_test/scgpt/test_resources/Kim2020_Lung.h5mu
+    - path: /resources_test/scgpt/source/vocab.json
 
 platforms:
   - type: docker
@@ -81,14 +77,10 @@ platforms:
     setup:
       - type: python
         __merge__: /src/base/requirements/anndata_mudata.yaml
-      - type: docker
-        run:
-          echo "build --jobs 4" >> /etc/bazel.bazelrc
       - type: python
         packages:
           - scgpt==0.2.1
     test_setup:
       - type: python
-        __merge__: [ /src/base/requirements/viashpy.yaml ]
     __merge__: [ /src/base/requirements/python_test_setup.yaml, .]
   - type: nextflow
